Improve LSP lifecycle UI tracing (#148)

+++ ui/src/windows/editor/editor-open-file-item/editor-open-file-item.component.ts
@@ -3,58 +3,87 @@ import {
   computed,
   inject,
   input,
-  OnDestroy,
   OnInit,
   signal,
   Signal,
-} from '@angular/core';
-import { MatButtonModule } from '@angular/material/button';
-import { MatIconModule } from '@angular/material/icon';
-import { EditorStateService } from '../core/state/editor-state.service';
-import { fileNode, voidCallback } from '../../../gen/type';
-import { MatTooltipModule } from '@angular/material/tooltip';
-import { EditorFileOpenerService } from '../core/services/editor-file-opener.service';
-import { removeFileNodeIfExists } from '../../../shared/file-node-helpers';
-import { ApplicationConfirmationService } from '../../../shared/services/application-confirmation.service';
-import { EditorFileStateService } from '../core/services/editor-file-state.service';
-import { EditorSessionStateService } from '../core/services/editor-session-state.service';
+} from "@angular/core";
+import { MatButtonModule } from "@angular/material/button";
+import { MatIconModule } from "@angular/material/icon";
+import { EditorStateService } from "../core/state/editor-state.service";
+import { fileNode, voidCallback } from "../../../gen/type";
+import { MatTooltipModule } from "@angular/material/tooltip";
+import { EditorDocumentOpenerService } from "../core/services/editor-document-opener.service";
+import { removeFileNodeIfExists } from "../../../shared/file-node-helpers";
+import { ApplicationConfirmationService } from "../../../shared/services/application-confirmation.service";
+import { EditorDocumentStateService } from "../core/lsp/editor-document-state.service";
+import { EditorSessionStateService } from "../core/services/editor-session-state.service";
+import { EditorDocumentDiagnosticService } from "../core/lsp/editor-document-diagnostic.service";
+import { useEffect } from "../../../lib/useEffect";
+import { Diagnostic } from "vscode-languageserver-protocol";
+import { EditorLanguageServerProtocolService } from "../core/lsp/editor-language-server-protocol.service";
+import { EditorDocumentLanguageIdService } from "../core/lsp/editor-document-language-id.service";
+import { EditorDocumentOpenTrackerService } from "../core/lsp/editor-document-open-tracker.service";
 
 @Component({
-  selector: 'app-editor-open-file-item',
+  selector: "app-editor-open-file-item",
   imports: [MatButtonModule, MatIconModule, MatTooltipModule],
-  templateUrl: './editor-open-file-item.component.html',
-  styleUrl: './editor-open-file-item.component.css',
+  templateUrl: "./editor-open-file-item.component.html",
+  styleUrl: "./editor-open-file-item.component.css",
 })
-export class EditorOpenFileItemComponent implements OnInit, OnDestroy {
+export class EditorOpenFileItemComponent implements OnInit {
   private readonly editorStateService = inject(EditorStateService);
-  private readonly editorFileOpenerService = inject(EditorFileOpenerService);
+  private readonly editorDocumentOpenerService = inject(
+    EditorDocumentOpenerService,
+  );
   private readonly applicationConfirmationService = inject(
     ApplicationConfirmationService,
   );
-  private readonly editorFileStateService = inject(EditorFileStateService);
-  private readonly editorSessionStateService = inject(EditorSessionStateService)
+  private readonly editorDocumentStateService = inject(
+    EditorDocumentStateService,
+  );
+  private readonly editorSessionStateService = inject(
+    EditorSessionStateService,
+  );
+  private readonly editorDocumentDiagnosticService = inject(
+    EditorDocumentDiagnosticService,
+  );
+  private readonly editorLanguageServerProtocolService = inject(
+    EditorLanguageServerProtocolService,
+  );
+  private readonly editorDocumentLanguageIdService = inject(
+    EditorDocumentLanguageIdService,
+  );
+  private readonly editorDocumentOpenTrackerService = inject(
+    EditorDocumentOpenTrackerService,
+  );
 
-  private unsub: voidCallback | null = null;
+  /**
+   * Displays count of error diagnostic it has for file
+   * @param diagnostics List of diagnostics
+   */
+  private displayDocumentDiagnostics = (diagnostics: Diagnostic[]) => {
+    this.errorDiagnosticCount.set(diagnostics.length);
+  };
+
+  constructor() {
+    useEffect(() => {
+      this.displayDocumentDiagnostics(
+        this.editorDocumentDiagnosticService.getDiagnostics(
+          this.fileNode().path,
+        ),
+      );
+    }, [this.editorDocumentDiagnosticService.valueChanged]);
+
+    useEffect(() => {
+      this.isDirty.set(
+        this.editorDocumentStateService.isDirty(this.fileNode().path),
+      );
+    }, [this.editorDocumentStateService.dirtyChanged]);
+  }
 
   ngOnInit() {
     this.openFileTooltip.set(this.fileNode().path);
-
-    this.unsub = this.editorFileStateService.onDirtyChange(
-      this.fileNode().path,
-      (isDirty) => {
-        this.isDirty.set(isDirty);
-      },
-    );
   }
-
-  ngOnDestroy() {
-    this.unsub?.();
-  }
-
-  /**
-   * TODO: on close check if there are unsaved changes for file and not auto save on then display warning beofre closing and restting state
-   * if press save apply changes to a string then save them to the file
-   */
 
   /**
    * Input file node to render for the given item
@@ -64,7 +93,7 @@ export class EditorOpenFileItemComponent implements OnInit, OnDestroy {
   /**
    * Holds the tooltip for hover information
    */
-  public openFileTooltip = signal('');
+  public openFileTooltip = signal("");
 
   /**
    * How long it takes for the parent tooltip to show
@@ -77,6 +106,13 @@ export class EditorOpenFileItemComponent implements OnInit, OnDestroy {
   public readonly isDirty = signal(false);
 
   /**
+   * Keep track selected directory
+   */
+  private readonly workspaceFolder = computed(() =>
+    this.editorStateService.selectedDirectoryPath(),
+  );
+
+  /**
    * Keep track if the given file tab is the one open / active
    */
   public isActive: Signal<boolean> = computed(
@@ -86,29 +122,34 @@ export class EditorOpenFileItemComponent implements OnInit, OnDestroy {
   );
 
   /**
+   * Holds how many error diagnostics this document has
+   */
+  public readonly errorDiagnosticCount = signal(0);
+
+  /**
    * The icon name displayed for the given file based on it's extension computed once
    */
   public fileIcon: Signal<string> = computed(() => {
     return (
       this.fileIconListMapNames.find(
         (x) => x.fileExtension == this.fileNode().extension,
-      )?.iconName ?? 'description'
+      )?.iconName ?? "description"
     );
   });
 
   private fileIconListMapNames: { fileExtension: string; iconName: string }[] =
     [
       {
-        fileExtension: '.html',
-        iconName: 'html',
+        fileExtension: ".html",
+        iconName: "html",
       },
       {
-        fileExtension: '.css',
-        iconName: 'css',
+        fileExtension: ".css",
+        iconName: "css",
       },
       {
-        fileExtension: '.js',
-        iconName: 'javascript',
+        fileExtension: ".js",
+        iconName: "javascript",
       },
     ];
 
@@ -120,15 +161,34 @@ export class EditorOpenFileItemComponent implements OnInit, OnDestroy {
 
     if (this.isDirty()) {
       const confirmed = await this.applicationConfirmationService.request(
-        'This file has unsaved changes are you sure you want to close it',
+        "This file has unsaved changes are you sure you want to close it",
       );
       if (!confirmed) {
         return;
       }
     }
 
-    this.editorFileStateService.reset(this.fileNode().path);
-    this.editorSessionStateService.removeCache(this.fileNode().path)
+    const workspaceFolder = this.workspaceFolder();
+    if (!workspaceFolder) {
+      console.error("Could not get workspace folder to close document");
+      return;
+    }
+
+    this.editorDocumentStateService.reset(this.fileNode().path);
+    this.editorSessionStateService.removeCache(this.fileNode().path);
+    this.editorDocumentOpenTrackerService.close(this.fileNode().path);
+
+    const languageId = this.editorDocumentLanguageIdService.getLanguageId(
+      this.fileNode().path,
+    );
+    if (languageId) {
+      this.editorLanguageServerProtocolService.didCloseTextDocument(
+        workspaceFolder,
+        languageId,
+        this.fileNode().path,
+      );
+      console.log("Send text document closed");
+    }
 
     let openfiles = this.editorStateService.openFiles() ?? [];
     removeFileNodeIfExists(openfiles, this.fileNode());
@@ -138,7 +198,7 @@ export class EditorOpenFileItemComponent implements OnInit, OnDestroy {
     if (this.isActive()) {
       let nextAvNode: fileNode | null = openfiles[0];
       if (nextAvNode) {
-        this.editorFileOpenerService.openFileNodeInEditor(nextAvNode);
+        this.editorDocumentOpenerService.openFileNodeInEditor(nextAvNode);
       }
     }
   }
@@ -147,6 +207,6 @@ export class EditorOpenFileItemComponent implements OnInit, OnDestroy {
    * Selects the given tab item node as the new active
    */
   public selectFileTabItem(event: Event) {
-    this.editorFileOpenerService.openFileNodeInEditor(this.fileNode());
+    this.editorDocumentOpenerService.openFileNodeInEditor(this.fileNode());
   }
 }
